modify datatype and pyproject.toml

+++ pyvrp/Model.py
@@ -119,8 +119,8 @@ class Model:
 
     def add_client(
         self,
-        x: int,
-        y: int,
+        x: float,
+        y: float,
         delivery: int = 0,
         pickup: int = 0,
         service_duration: int = 0,
@@ -189,8 +189,8 @@ class Model:
 
     def add_depot(
         self,
-        x: int,
-        y: int,
+        x: float,
+        y: float,
         tw_early: int = 0,
         tw_late: int = np.iinfo(np.int64).max,
         *,
