+++ pyvrp/_pyvrp.pyi
@@ -29,8 +29,8 @@ class DynamicBitset:
     def reset(self) -> DynamicBitset: ...
 
 class Client:
-    x: int
-    y: int
+    x: float
+    y: float
     delivery: int
     pickup: int
     service_duration: int
@@ -43,8 +43,8 @@ class Client:
     name: str
     def __init__(
         self,
-        x: int,
-        y: int,
+        x: float,
+        y: float,
         delivery: int = 0,
         pickup: int = 0,
         service_duration: int = 0,
@@ -74,15 +74,15 @@ class ClientGroup:
     def clear(self) -> None: ...
 
 class Depot:
-    x: int
-    y: int
+    x: float
+    y: float
     tw_early: int
     tw_late: int
     name: str
     def __init__(
         self,
-        x: int,
-        y: int,
+        x: float,
+        y: float,
         tw_early: int = 0,
         tw_late: int = ...,
         *,
